new loss function with Huber

+++ modules/sctp/sctp/utils/underlying_graph.py
@@ -23,6 +23,7 @@ class ProbabilisticGraph:
     probabilities: np.ndarray
 
 def get_initial_edges(graph):
+    # edges with 0-based indices for the underlying graph
     initial_edges = []
     for poi in graph.pois:
         initial_edges.append([poi.neighbors[0]-1, poi.neighbors[1]-1, poi.block_prob])
@@ -60,12 +61,12 @@ def create_adj_prob_matrices(edges, positions):
         (n, n) symmetric matrix where entry (i,j) is the Euclidean distance
         between vertices i and j if <= threshold, else 0.
     """
-    adjacency = np.zeros((positions.shape[0], positions.shape[0]))
-    probabilities = np.zeros((positions.shape[0], positions.shape[0]))
+    adjacency = np.zeros((positions.shape[0], positions.shape[0]), dtype=np.float32)
+    probabilities = np.zeros((positions.shape[0], positions.shape[0]), dtype=np.float32)
     # Compute upper triangle only
     for edge in edges:
         dist = np.linalg.norm(np.array(positions[edge[0]]) - np.array(positions[edge[1]]))
-        assert adjacency[edge[0], edge[1]] == 0
+        assert adjacency[edge[0], edge[1]] == 0.0
         adjacency[edge[0], edge[1]] = dist
         probabilities[edge[0], edge[1]] = edge[2]
